% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_cleanup.R
\name{calculate_vintage_hashes}
\alias{calculate_vintage_hashes}
\title{Calculate vintage hashes}
\usage{
calculate_vintage_hashes(vintage_ids, con, schema = "platform")
}
\arguments{
\item{vintage_ids}{vector of vintage ids}

\item{con}{connection to database}

\item{schema}{schema name, defaults to "platform"}
}
\value{
dataframe with vintage id and both hashes
}
\description{
Calculates the hash of all the datapoints in a particular vintage as well
as a partial hash, which excludes the last datapoint, allowing the hashes
of sucessive vintages to be easily compared.
}
