% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{common_parameters}
\alias{common_parameters}
\title{Parameter definitions for common parameters}
\arguments{
\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{schema}{character schema name, default is "platform"}

\item{code_no}{character table code}
}
\description{
Parameter definitions for common parameters
}
