% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ensure_colnames_utf8}
\alias{ensure_colnames_utf8}
\title{Ensure proper UTF-8 encoding for dataframe column names}
\usage{
ensure_colnames_utf8(df)
}
\arguments{
\item{df}{A dataframe to fix column names encoding}

\item{from}{Source encoding to try}
}
\value{
A dataframe with valid UTF-8 column names
}
\description{
Ensure proper UTF-8 encoding for dataframe column names
}
