% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_category_table}
\alias{insert_new_category_table}
\title{Insert a new category-table link}
\usage{
insert_new_category_table(con, df, schema = "platform")
}
\arguments{
\item{con}{Database connection object}

\item{df}{A data frame with one row containing the following columns:
* category_id (integer): ID of the category
* table_id (integer): ID of the table
* source_id (integer): ID of the source these both belong to}

\item{schema}{Character string specifying the database schema. Defaults to "platform"}
}
\value{
A data frame with one column 'count' indicating number of rows inserted
        (1 for success, 0 if link already exists)
}
\description{
Links a table to a category within the same source. A table can be linked
to multiple categories.
}
\examples{
\dontrun{
df <- data.frame(
  category_id = 1,
  table_id = 1,
  source_id = 1
)
insert_new_category_table(con, df)
}

}
