% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_data}
\alias{insert_new_data}
\title{Insert new data for a table i.e. a vintage}
\usage{
insert_new_data(code_no, con, schema = "platform")
}
\arguments{
\item{code_no}{character table code}

\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{schema}{character schema name, default is "platform"}
}
\value{
list of tables with counts for each inserted row.
}
\description{
When new data for a table (in SURS speak 'matrix") is added, these are new
vintages. This function adds a set of new vintages and their corresponding
data points and flags to the database, by calling the respective SQL functions
for each of these tables.
}
\examples{
\dontrun{
purrr::walk(master_list_surs$code, ~insert_new_data(.x, con))
}
}
