% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_writing.R
\name{insert_new_unit}
\alias{insert_new_unit}
\title{Insert a new unit}
\usage{
insert_new_unit(con, df, schema = "platform")
}
\arguments{
\item{con}{Database connection object}

\item{df}{A data frame with one row containing the following columns:
* name (character): Name of the unit}

\item{schema}{Character string specifying the database schema. Defaults to "platform"}
}
\value{
A data frame with one column 'count' indicating number of rows inserted
        (1 for success, 0 if unit with same name already exists)
}
\description{
Inserts a new unit into the unit table. If a unit with the same name
already exists, the function will not insert a duplicate and will return 0.
}
\examples{
\dontrun{
df <- data.frame(
  name = "meters"
)
insert_new_unit(con, df)
}

}
