% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintage_cleanup.R
\name{process_no_keep_vintage_table}
\alias{process_no_keep_vintage_table}
\title{Process a table with keep_vintage = FALSE}
\usage{
process_no_keep_vintage_table(con, schema, table_id, result)
}
\arguments{
\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema}

\item{table_id}{Integer specifying the table to process}

\item{result}{List to store results}
}
\value{
Updated result list
}
\description{
Process a table with keep_vintage = FALSE
}
\keyword{internal}
