% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_deleting.R
\name{remove_empty_vintages}
\alias{remove_empty_vintages}
\title{Clean up function for leftover vintages without datapoints}
\usage{
remove_empty_vintages(con, schema = "platform")
}
\arguments{
\item{con}{Database connection object}

\item{schema}{Character string specifying the database schema. Defaults to "platform"}
}
\value{
invisibly the deleted vintage ids
}
\description{
cleans up vintages without data points. Also cascades to flags.
}
