% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sql_function_call}
\alias{sql_function_call}
\title{Construct and execute an SQL function call}
\usage{
sql_function_call(con, fun_name, args, schema = "platform")
}
\arguments{
\item{con}{PostgreSQL connection object created by the RPostgres package.}

\item{fun_name}{character name of SQL function to call}

\item{args}{list of arguments, can be named (but then all of them have to be)}

\item{schema}{character schema name, default is "platform"}
}
\value{
value of `dbGetQuery(con, "SELECT * FROM schema.fun_name($args)")$fun_name`
}
\description{
Constructs the call for the funciton `schema`.`fun_name` from the database
with the arguments `args` and returns the result. Because of the way
the parameters get apssed the hash of the query remains the same which is
a bitch for testing with mock fixtures using dittodb. So we added a comment
to the query, just to ensure the hash is different.
}
\details{
Inspiration from `timeseriesdb` package by Matt Bannert.
}
